Remove gracenote SDK
