+++ app/src/main/java/com/schober/vinylcast/recognizer/MusicRecognizer.java
@@ -1,0 +1,6 @@
+package com.schober.vinylcast.recognizer;
+
+public interface MusicRecognizer {
+    void start();
+    void stop();
+}
